requirements.txt
matplotlib
jax
numpy
pandas
scipy
astropy
spectral-cube
astroquery
astroML
paperswithcode-client
python-dotenv